Bug 567032 issue floating license configuration wizard
 Rework existing LicenseOperatorService to extend it with floating
license issuing.
