+++ bundles/org.eclipse.passage.lic.floating.model/META-INF/MANIFEST.MF
@@ -12,15 +12,10 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0";x-installation:=
  org.eclipse.emf.ecore;bundle-version="0.0.0";visibility:=reexport,
  org.eclipse.passage.lic.emf.meta;bundle-version="0.6.0",
  org.eclipse.passage.lic.floating;bundle-version="0.0.0";visibility:=reexport
-Export-Package: org.eclipse.passage.lic.floating.model.api;
-  x-friends:="org.eclipse.passage.lic.floating.edit",
+Export-Package: org.eclipse.passage.lic.floating.model.api;x-friends:="org.eclipse.passage.lic.floating.edit,org.eclipse.passage.loc.dashboard.ui,org.eclipse.passage.loc.api",
  org.eclipse.passage.lic.floating.model.impl;x-internal:=true,
- org.eclipse.passage.lic.floating.model.meta;
-  x-friends:="org.eclipse.passage.lic.floating.e4.ui,
-   org.eclipse.passage.lic.floating.edit",
- org.eclipse.passage.lic.floating.model.util;
-  x-friends:="org.eclipse.passage.lic.floating.edit",
- org.eclipse.passage.lic.internal.floating.model;
-  x-friends:="org.eclipse.passage.loc.dashboard.ui"
+ org.eclipse.passage.lic.floating.model.meta;x-friends:="org.eclipse.passage.lic.floating.e4.ui,org.eclipse.passage.lic.floating.edit",
+ org.eclipse.passage.lic.floating.model.util;x-friends:="org.eclipse.passage.lic.floating.edit",
+ org.eclipse.passage.lic.internal.floating.model;x-friends:="org.eclipse.passage.loc.dashboard.ui"
 Bundle-ClassPath: .
 Bundle-ActivationPolicy: lazy
