+++ bundles/org.eclipse.passage.loc.api/META-INF/MANIFEST.MF
@@ -26,4 +26,5 @@ Require-Bundle: org.eclipse.e4.core.services;bundle-version="0.0.0",
  org.eclipse.passage.lic.licenses;bundle-version="1.0.0";visibility:=reexport,
  org.eclipse.passage.lic.products;bundle-version="1.0.0";visibility:=reexport,
  org.eclipse.passage.lic.users;bundle-version="1.0.0";visibility:=reexport,
- org.eclipse.passage.lic.users.model;bundle-version="1.0.0"
+ org.eclipse.passage.lic.users.model;bundle-version="1.0.0",
+ org.eclipse.passage.lic.floating.model;bundle-version="0.1.0"
